requirements.txt
tensorflow
numpy
pandas
matplotlib
mlflow
scikit-learn
pillow
tqdm
fastapi
uvicorn
boto3
